requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
python-dateutil>=2.9.0
# Optional Firestore support
firebase-admin>=6.4.0
